requirements.txt
googletrans==4.0.0rc1
HF_API==0.1
hfapi==0.1b0
interactions_files==1.1.3
jieba==0.42.1
pymongo==4.2.0
python_telegram_bot==13.13
requests==2.28.1
telegram==0.0.1
toml==0.10.2
